feat: add Dockerfile and CI workflow (#1)
This adds a dockerfile with all the (currently known) dependencies added for speedy CI, as well as a CI workflow that will eventually test code format and build the kernel. Once we have a zig build file, we'll want to add the format checking, building and testing steps to the workflow.